# web-bff/src/swaggerOutput.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Manage users",
    "description": "FreshLink manage users service"
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": ""
    }
  ],
  "paths": {
...
    "/farmers/": {
      "get": {
        "tags": [
          "Farmers"
        ],
        "summary": "Get all farmers",
        "description": "Returns a list of all farmers",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with a list of farmers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/farmer"
                  }
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/arrayOfFarmers"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Farmers"
        ],
        "summary": "Register a new farmer",
        "description": "Adds a new farmer",
        "responses": {
          "200": {
            "description": "Farmer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/farmer"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/farmer"
                  }
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Already exists": {
                    "value": {
                      "message": "Email already exists"
                    }
                  }
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Missing required properties": {
                    "value": {
                      "message": "All required properties must be defined"
                    }
                  }
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/hiddenSchemas/postFarmer"
              },
              "examples": {
                "example1": {
                  "$ref": "#/components/examples/postFarmer"
                }
              }
            }
          }
        }
      }
    },
    "/farmers/{id}": {
      "get": {
        "tags": [
          "Farmers"
        ],
        "summary": "Get farmer by id",
        "description": "Returns farmer object with specified id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Farmer ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with a farmer object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/farmer"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/farmer"
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Invalid ID format": {
                    "value": {
                      "message": "Id is incorrect"
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Object not found": {
                    "value": {
                      "message": "Object with id '677164acd5343f0dac83716a' does not exist"
                    }
                  }
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Farmers"
        ],
        "summary": "Update a farmer",
        "description": "Updates an existing farmer's details. Email can not be changed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Farmer ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Farmer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/farmer"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/farmer"
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Invalid ID format": {
                    "value": {
                      "message": "Id is incorrect"
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Object not found": {
                    "value": {
                      "message": "Object with id '677164acd5343f0dac83716a' does not exist"
                    }
                  }
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Missing required properties": {
                    "value": {
                      "message": "All required properties must be defined"
                    }
                  }
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/hiddenSchemas/farmerWithoutId"
              },
              "examples": {
                "example1": {
                  "$ref": "#/components/examples/farmerWithoutId"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Farmers"
        ],
        "summary": "Delete a farmer",
        "description": "Deletes a farmer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Farmer ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Farmer successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully deleted"
                    }
                  }
                },
                "examples": {
                  "Success": {
                    "value": {
                      "message": "Successfully deleted"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Invalid ID format": {
                    "value": {
                      "message": "Id is incorrect"
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "No token": {
                    "value": {
                      "message": "No token provided"
                    }
                  },
                  "JWT payload": {
                    "value": {
                      "message": "Cannot parse JWT payload"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Bad token": {
                    "value": {
                      "message": "Invalid or expired token"
                    }
                  },
                  "Role": {
                    "value": {
                      "message": "Your role does not have access"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Object not found": {
                    "value": {
                      "message": "Object with id '677164acd5343f0dac83716a' does not exist"
                    }
                  }
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/farmers/login": {
      "post": {
        "tags": [
          "Farmers"
        ],
        "summary": "Login a farmer",
        "description": "Authenticates a farmer and returns a JWT token and farmer details",
        "responses": {
          "200": {
            "description": "Farmer logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "your-jwt-token"
                    },
                    "farmer": {
                      "$ref": "#/components/schemas/farmer"
                    }
                  }
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/loginFarmerResponse"
                  }
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Invalid credentials": {
                    "value": {
                      "message": "Invalid email or password"
                    }
                  }
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "Missing required properties": {
                    "value": {
                      "message": "All required properties must be defined"
                    }
                  }
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                },
                "examples": {
                  "example1": {
                    "$ref": "#/components/examples/error500"
                  }
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "example1": {
                  "summary": "A sample farmer login",
                  "value": {
                    "email": "john.doe@example.com",
                    "password": "password123"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Get the token from login endpoints"
      }
    },
    "schemas": {
...
      "farmer": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "677164acd5343f0dac83716a"
                  }
                }
              },
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "john.doe@example.com"
                  }
                }
              },
              "farmName": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Doe's Farm"
                  }
                }
              },
              "description": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "A nice farm"
                  }
                }
              },
              "name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "John"
                  }
                }
              },
              "surname": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Doe"
                  }
                }
              },
              "location": {
                "$ref": "#/components/schemas/location"
              }
            }
          }
        }
      },
...
    },
...
    "examples": {
...
      "error500": {
        "summary": "Server error",
        "value": {
          "message": "Internal Server Error"
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}